Add the reset_follower management command
Add a new Django management command, `reset_follower`, that can
manually reset the tracking states of a follower (i.e. a
ProcessApplication instance).  The idea for the follower is to
replay the whole history next time it is (manually) synchronised.

Note that this command does not interact with a running system,
see the command `sync_followers` to nudge the system in catching
up with the upstream apps.